Auction Query Rewrites + More
Auction query class has basically been rewritten in areas to support better flow for multiple types of queries, and to support better handling of new retail AH Throttle Dropped Events etc.

There are now auctioning operations. Still missing some options, such as specifiying what happens if below min price or above max price etc. Also missing options for cancelling. Will be added later. But the basis is there. It is not quite available fully on Classic yet. As I have not finished implementing the UI for classic. However, the underlying code for the AH interaction is there. On retail if you are using the base auction house, and go to the sell or auctions tab. You will now be presented with a side list to your right that lists the available auction operations with one button for post scan and one for cancel scan.

Eventually plan to add Macro settings for binding scroll wheel for the various actions, which will incorporate not only Ans, but also allow direct usage of TSM buttons as well. So a one stop shop macro if you use both Ans and TSM together.

Sniper slightly rewritten to take advantage of the new query class changes. Not fully tested on Classic yet, so use this update at your own risk on classic. Has been tested on retail.

Changed event manager emit function slightly, so I could get proper error tracing reported by WoW Client.

Fixed an issue where the group name was not updating in the tree view.
Groups now can be dragged and dropped onto other groups for easier group parenting. Up and down buttons still required for ordering

Moved some general methods out of Sniping operation and into Utils for use in other operations.

Fixed an error in DashboardView where a variable was not declared properly.
Fixed an error in LedgerView where a variable was not declared properly.
Fixed an issue in LedgerView where the default sort for time was Ascending instead of Descending.