Vendor Buy Data Added + more
can now use vendorbuy in stuff to get what the item may be price wise if a vendor sells it.

Fixed a bug in query where the statements were flipped.

Added new Skip Lowest for groups in sniper settings. It will skip a group that it has already seen the lowest on, until a new low is found.

Added chat messages for new items in classic for sniper settings.

New method for buying stuff for the sniper. Previously, the sniper would try and buy the item based on the index of the selected row when double clicked. Now, the selected item is stored based on the last selected row, and a new button is shown to the right to let you know how much it will be total, what the item is, etc. Clicking the button will try and buy that selected item. Double clicking has been removed. However, the AuctionSnipe:BuyFirst() and AuctionSnipe:BuySelected() still work for macros.

Removed an unneeded test binding.

Owner is now listed again on the sniper. For retail if there are mutliple it just says Multiple. However, if it is an individual it will list their name or blank if not available yet.