Several Bug Fixes for Classic
Fixed an issue where Auctioning operation would not  post the next item on classic.

Fixed an issue where classic would fail to buy grouped items on the sniper.

Fixed an issue where when adding chat message and auction owner was not ready yet, it would throw an error.

Added back in double click purchasing for classic sniper.

Fixed an issue where upon closing the sniper tab, items were not cleared on classic properly.

Fixed a nil error if sorting by owner on both classic and retail.

Fixed an issue where an improper background asset was trying to be used for classic.

Fixed an issue where the size of the background graphic for the new buy button on sniper for classic was incorrectly sized.

Fixed an issue where text would not truncate properly on both retail and classic for the new buy button.

Fixed an issue where the classic sniper header had moved too far to the left.