Implemented a workaround for purchasing auctions during a browse state for sniper
Apparently, the internal auction house in the client expects the client to be in a SEARCH item state to be able to send PlaceBid. Otherwise, the client just does nothing. It doesn't matter if the search state is for the auction id being bought. The wow client just has to be in such a state that a Search Item query has been sent in some form.