Moved to Finite State Machines
Moved to Finite State Machines For Query, Sniper, and Post / Cancel.

Updated BagScanner with more info and utils.

Post / Cancel scan now have additional checks for verification of posting the proper item, and verifying the item was cancelled successfully. Post / Cancel scan no  longer fail in certain cases.

Added a debug log window. By default it is not written to, unless the it is active. Can turn the window on from Settings -> General

In formulas, you now have access to bonus() which returns the 2 parameter if it has the specified item bonus. Otherwise it will return the 3rd parameter if it does not have the specified item bonus. The first parameter is the item bonus id as a number. Thus: bonus(14445, 10000g, 0c)

eq, neq, startsWith, and contains, now have will return the 3rd parameter if true, otherwise will return the 4th parameter if false. E.g.: eq(ilevel, 400, 10000g, 0c)

Corrected text for auction hours on classic, it was 8,12,24. when it should have been 2,8,24 hours.

Sniper now highlights the latest found auctions between Browse / Browse More scans on Retail. It will unhighlight them on the next browse / browse more scan.

Sniper now confirms item purchases in the following cases on Retail: A browse query was previous executed, an item search query was sent but not complete, or sniper was stopped but the previous query was a browse query. Otherwise, if it is during the scan delay of Waiting to Query the item will be purchased without a confirm. This only applies to Retail. Classic is unchanged.

Both classic and retail now have double click row to buy again on sniper.

Corrected an issue where the wrong row was highlighted on Retail in some cases.

Fixed an issue on Classic where it was removing the wrong auction from an array, thus not allowing stacked auctions to be bought after the first one.

Moved to Shift + Ctrl + Click for blacklisting item ids. Shift + Click to temporary blacklist.