Cleared Data.lua and Fixed another issue
Fixed an issue where Query.SearchForItem was being called on the current sniper scan index without cloning it for SearchForItem. Thus, it could mess up on scans if you bought something and it had to rescan the previous scan index for a battle pet. Regular items and commodities were unaffected, as the comparison fields for certain checks were not changed. They were only changed if it was a battle pet.