Added support for i:id(ilevel) item db lookup
Now saves sniper config for the sniper window: clevel range, min quality, and min ilevel.

Added option to toggle recalc of percentage for sniping operations. Recalc will only apply if you set a Max Price formula and not a Boolean formula for the operation and Recalc Percentage is checked.

Added icon flash for wow client when items are found on sniping.

Reworked last seen items for sniper and it now removes items that are no longer there when a search is done again. This only applies to Retail.

Fixed a bug where it would go into the buy state on sniper via scroll wheel macro, even if no item was selected.