Fixed Several Issues on Classic
First issue fixed is the quality dropdown text not being properly set. I did not realize, blizz changed how SetSelectedID worked on retail in 8.x compared to Classic legion ui api of 7.x. Where in in 8.x SetSelectedID would properly update the dropdown text with the proper value. However, on classic, this is not the case and must be updated manually.

Fixed an issue where on Logger edit box it was trying to use an edit box widget api that is only available in 8.x and not in 7.x ui api. I just removed it completely for now. Also added in autoFocus="false" to prevent auto focusing the logger text area when it pops up. Renamed the names of the edit box plus scrollframe as well to be more unique.