2.2.7 Retail Beta Initial Changes
Underlying system rework for state machine / events etc.
Reworked snipe, post and cancel to use the new systems.

A gem icon is now shown for socketed items in sniper view.
Added a new variable for formulas: socket
The socket variable returns true if there is a socket on the item.

Fixed a bug where a variable was name incorrectly, and thus a string was never being cached properly. Causing extra unneeded computation time and memory gain.

Optimized some stuff in Sources.lua. Now requires manually adding the required variables to the template. This help prevent one extra computation of string formatting and one extra loop. Fixed an issue where a string was being created even all the time. The string was meant to only be created if the auction was a pet.

Added support for latest TSM group data format. It can now be imported properly into AnS.

The default config for itemsPerUpdate changed to 50 instead of 20.
Added another config item to check for classicMode for sniper. In order to help correct an issue where quality field was not reset properly if you launched retail and then classic.

The first() in formulas next except any number of arguments instead of just 2.

Fixed an issue where group import fails to import if the group being imported is the selected group.

Fixed an issue where the wrong font was used for traditional chinese, thus throwing an error on Classic WoW.

Added a clear auction data button for AnsAuctionData in Settings -> General
Reworked AnsAuctionData scanner to work on Retail as well using the new Replicate feature of retail.
Reworked AnsAuctionData tooltip positioning.
Fixed an issue where some auctions were never being tracked, because there were 5 of them or less.
Fixed an issue in statistics where it was returning 0,0,0,0 in some cases for: min,avg,sum,count. When it should have just been returning the previous known values.
Fixed an issue where a temp table was not being recycled properly.

On retail ItemKeyInfo is now cached for sniper, allowing support for more than 1000 that the client would only cache.