Bug Fixes & Classic Ready
Fixed blizzards retarded TBC UI Auction House error for them by overriding the AuctionFrameBrowse_Update and applying a check to prevent the issue.

Hopefully, fixed the issue in retail, where in the item is already purchased and should just be removed and skipped.

Fixed sort issue on sniper and it now properly remembers last sort.

Added in Ignore Item Button for sniper. Also added in the corresponding Macro creation to trigger this button as well.

Fixed an issue in AnsAuctionData scanning missing items.

Cleaned up AnsAuctionData code a little bit in regards to loading of data / getting data.

Added in a modal notification for sniper on retail to alert the user to having to pause sniper to purchase / item no longer available etc.

Finalized classic sniper changes and tested.

TODO: Classic and Retail sniper Ignore Item button are in reversed positions. Need to make this consistent.