Various Bug Fixes
Fixed issue with min quality for classic
Fixed throttling issue on classic hopefully
Fixed issue in blacklist item add, where when given just a number to add, the improper string format was used for the key in the dictionary.

Blacklist is now sorted alphanumerically based on link if there is one, otherwise by item id.

Fixed text overlap of checkboxes for macro settings

Increased destroy window width to help with text overlap

Fixed issue on retail, where in latest patch blizz changed the Enum for Qualities by settings none to 0 and increasing every other by 1. It now directly references Blizz Api of Enum.AuctionHouseFilter on retail.

Fixed issue where if Ignore First Macro was used, the list was not refreshed properly if the first item was not selected.