Several bug fixes + enhancements
AnS now interfaces automatically with TSM custom user defined data sources. The name is the same as used in TSM for access in AnS as well, except case is ignored in AnS.

Fixed issue where ItemTracker was tracking bound inventory items. Obviously this should not be the case at all in relation to Auctioning.

Fixed query sort issue on retail by creating a separate sort for commodities / items.

Fixed issue where previous registered sources could be overwritten when they shouldn't be. If a source tries to overwrite an existing one. The new one is now ignored.

Fixed sorting error issues for blacklist UI and sniper quality sort issue when upgrading from an older version of AnS.

Enhancement: If all items are purchased in sniper view on classic. It will now instantly start the next scan query rather than waiting as if there are still items available.