Wrapped C_AddOns
I have to wrap various C_AddOns functions into Utils in order to properly check whether the C_AddOns version of the function exists or not. More than likely once Cata Classic releases this won't be an issue and the API will once again be in a semi-synced state between Classic, Cata Classic and Retail.